Refractored SwaggerRoutes logic to SwaggerController
